% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_features.R
\name{count_features}
\alias{count_features}
\title{Assigning reads to genomic features}
\usage{
count_features(
  bam_dir = ".",
  annotation_dir = ".",
  output_dir = ".",
  annotation_file,
  chromosome_alias_file,
  target_features,
  strandedness,
  is_paired_end,
  ...
)
}
\arguments{
\item{bam_dir}{The directory where bam files can be found.}

\item{annotation_dir}{The directory where the annotation file can be found.}

\item{output_dir}{The directory where the resulting CSV files will be deposited.}

\item{annotation_file}{A GFF3 or GTF genome annotation file.}

\item{chromosome_alias_file}{A comma-delimited TXT file containing a character string with the chromosome names. This file has to have two columns: first with the chromosome name in the annotation file, second with the chromosome name in the BAM file.}

\item{target_features}{A vector of strings outlining all feature types to be examined.}

\item{strandedness}{A string outlining the type of the sequencing library: unstranded, stranded, or reversely stranded.}

\item{is_paired_end}{A boolean indicating if the reads are paired-end.}

\item{...}{Optional parameters passed on to featureCounts()}
}
\value{
Count tables for each feature are written into separate files, as well as the result summary.
}
\description{
A wrapper function allowing to count reads to different typed of genomic features in one go. Designed to incorporate newly-discovered sRNAs and UTRs into the analysis.
}
