% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{differential_expression}
\alias{differential_expression}
\title{Differential gene expression analysis}
\usage{
differential_expression(feature_count_file, metadata_file, cutoff_value,
  multiple_variables, main_condition, output_file_name)
}
\arguments{
\item{feature_count_file}{A CSV file containing a table of counts for a particular feature type.}

\item{metadata_file}{A file outlining the experimental conditions for each sample.}

\item{cutoff_value}{Retain only the genes that have at least half the samples with the counts above the cut-off value.}

\item{multiple_variables}{A boolean indicating if the experimental design contained more than one condition.}

\item{main_condition}{A string indicating which of the conditions is the most important; should coincide with the name in the metadata table.}

\item{output_file_name}{A string containing the name of the output file.}
}
\value{
DESeq result table and the result summary are written into separate files. Summary file name is derived from the provided output file name with an addition of "summary.txt".
}
\description{
A wrapper function to perform analysis of differential gene expression using DESeq2 method.
}
