% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_features.R
\name{find_strandedness}
\alias{find_strandedness}
\title{find_strandedness function}
\usage{
find_strandedness(strand_param)
}
\arguments{
\item{strand_param}{user input 'stranded' or 'reversely-stranded'}
}
\value{
strand_sp integer value
}
\description{
This function translates the user-inputted strandedness parameter to the required integer input for strandSpecific arg of featureCounts.
}
