% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_features.R
\name{make_saf}
\alias{make_saf}
\title{SAF converter}
\usage{
make_saf(ann_file, exclude = F)
}
\arguments{
\item{ann_file}{A GFF3 annotation file.}

\item{exclude}{A boolean to indicate whether or not to include rRNA/tRNA features}
}
\value{
A dataframe in Simplified Annotation Format
}
\description{
This function converts gff file into simplified annotation format with 5 columns, appropriate for use in featureCounts/tpm_norm_flagging.
}
