% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_features.R
\name{read_annotation_file}
\alias{read_annotation_file}
\title{read_annotation_file function}
\usage{
read_annotation_file(annot_dir, annot_file)
}
\arguments{
\item{annot_dir}{Directory where genome annotation file is located}

\item{annot_file}{GFF3 or GTF genome annotation file}
}
\value{
annot_file_loc Complete path of existing annotation file
}
\description{
This function pastes together the path and filename of annotation file (if not in current directory) and tests file for existence.
}
