% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_file_editor.R
\name{sRNA_calc}
\alias{sRNA_calc}
\title{sRNA prediction}
\usage{
sRNA_calc(major_strand_features, target_strand, union_peak_ranges)
}
\arguments{
\item{major_strand_features}{A dataframe containing the major features for a particular strand.}

\item{target_strand}{A character string indicating the strand. Supports two valies; '+' and '-'.}

\item{union_peak_ranges}{An IRannges object containing genomic coordinated for all peaks detected on the target strand.}
}
\value{
An IRanges object containing coordinates and names of the predicted sRNA.
}
\description{
The function for prediction and annotation of sRNA.
}
