% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_file_editor.R
\name{strand_feature_editor}
\alias{strand_feature_editor}
\title{Strand annotation}
\usage{
strand_feature_editor(
  target_strand,
  sRNA_IRanges,
  UTR_IRanges,
  major_strand_features
)
}
\arguments{
\item{target_strand}{A character string indicating the strand. Supports two values; '+' and '-'.}

\item{sRNA_IRanges}{An IRanges object containing coordinates and names of the predicted sRNAs.}

\item{UTR_IRanges}{An IRanges object containing coordinates and names of the predicted UTRs.}

\item{major_strand_features}{A dataframe containing the major features for a particular strand.}
}
\value{
A dataframe containing strand annotation populated with the prediction features, build in accordance with GFF3 file format.
}
\description{
This function constructs the full annotation for the strand.
}
