% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpm_norm_flagging.R
\name{tpm_flag_filtering}
\alias{tpm_flag_filtering}
\title{Filtering features by a target flag.}
\usage{
tpm_flag_filtering(
  complete_annotation_file,
  target_features,
  target_flag,
  output_file
)
}
\arguments{
\item{complete_annotation_file}{A flagged GFF3 annotation file.}

\item{target_features}{A string indicating feature type to filter.}

\item{target_flag}{A string indicating a flag for filtering.}

\item{output_file}{A string indicating the name of athe output file.}
}
\value{
A GFF3 file where a target feature type is filtered by the expression flag of interest.
}
\description{
A function to filter the marked features selected by the user by the flag of choice.
}
