% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpm_norm_flagging.R
\name{tpm_flagging}
\alias{tpm_flagging}
\title{Flagging features depending on TPM value profile}
\usage{
tpm_flagging(tpm_data, complete_annotation, output_file)
}
\arguments{
\item{tpm_data}{A CSV file containing TPM values for each feature in each sample.}

\item{complete_annotation}{A GFF3 annotation file.}

\item{output_file}{A string indicating the name of athe output file.}
}
\value{
A Gff3 file, where a target feature type has an expression flad added to its attribute column.
}
\description{
A helper function to analyse each row of the TPM table. Each feature gets allocated a flag depending on the expression profile.
}
