% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpm_norm_flagging.R
\name{tpm_normalisation}
\alias{tpm_normalisation}
\title{TPM calculator}
\usage{
tpm_normalisation(
  count_table,
  complete_ann,
  feature_type = c("putative_sRNA", "putative_UTR"),
  is_gff = T,
  output_file = NA,
  excl_rna = T
)
}
\arguments{
\item{count_table}{A CSV file containing feature counts for each sample.}

\item{complete_ann}{A GFF3 annotation file or SAF dataframe}

\item{feature_type}{A string indicating desired feature type(s) from annotation.}

\item{is_gff}{A boolean indicating whether annotation is gff file, default=T}

\item{output_file}{A string indicating the name of the output file.}

\item{excl_rna}{A boolean indicating if misc RNA features (rRNA, tRNA) are excluded from normalisation. (Defaults=T)}
}
\value{
A dataframe with TPM values for each gene and sample; the same is written into the output file.
}
\description{
This function uses feature count tables to calculate TPM values for each gene and sample.
}
