% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_file_editor.R
\name{UTR_calc}
\alias{UTR_calc}
\title{UTR prediction}
\usage{
UTR_calc(major_strand_features, target_strand, union_peak_ranges,
  min_UTR_length)
}
\arguments{
\item{major_strand_features}{A dataframe containing the major features for a particular strand.}

\item{target_strand}{A character string indicating the strand. Supports two valies; '+' and '-'.}

\item{union_peak_ranges}{An IRannges object containing genomic coordinated for all peaks detected on the target strand.}

\item{min_UTR_length}{An integer indicating the minimum UTR length.}
}
\value{
An IRanges object containing coordinates and names of the predicted UTRs.
}
\description{
The function for prediction and annotation of UTRs.
}
