% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_file_editor.R
\name{peak_union_calc}
\alias{peak_union_calc}
\title{Peak union calculation}
\usage{
peak_union_calc(bam_location = ".", target_strand, low_coverage_cutoff,
  high_coverage_cutoff, peak_width, paired_end_data = FALSE,
  strandedness = "unstranded")
}
\arguments{
\item{bam_location}{The directory containing BAM files.}

\item{target_strand}{A character string indicating the strand. Supports two valies; '+' and '-'.}

\item{low_coverage_cutoff}{An interger indicating the low coverage threshold value.}

\item{high_coverage_cutoff}{An interger indicating the high coverage threshold value.}

\item{peak_width}{An interger indicating the minimum peak width.}

\item{paired_end_data}{A boolean indicating if the reads are paired-end.}

\item{strandedness}{A string outlining the type of the sequencing library: stranded, or reversely stranded.}
}
\value{
An object of IRanges class, containing the genomic coordinates of selected and unified peaks.
}
\description{
The function goes over each BAM file in the directory and finds the expression peaks that satisfy the coverage boundary and length criteria in each file. Then it unifies the peak information to obtain a single set of peak genomic coordinates.
}
