% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpm_norm_flagging.R
\name{tpm_normalisation}
\alias{tpm_normalisation}
\title{TPM calculator}
\usage{
tpm_normalisation(count_table, complete_gff, target_feature, output_file = NA)
}
\arguments{
\item{count_table}{A CSV file countaining feature counts for each sample.}

\item{complete_gff}{A GFF3 annotation file.}

\item{target_feature}{A string containing the feature type.}

\item{output_file}{A string indicating the name of athe output file.}
}
\value{
A dataframe with TPM values for each gene and sample; the same is written into the output file.
}
\description{
This function uses feature count tables to calculate TPM values for each gene and sample.
}
